MediaAttachmentsの表示機能をコンポーネントに分離
